backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
torch>=2.0.0
torchaudio>=2.0.0
openai-whisper>=20231117
numpy>=1.24.0
soundfile>=0.12.1
